+++ buybackprogram/notification.py
@@ -1,6 +1,8 @@
 """
 notifications helper
 """
+import requests
+
 from django.contrib.auth.models import User
 
 from allianceauth.notifications import notify
@@ -119,54 +121,45 @@ def send_user_notification(user: User, level: str, message: dict) -> None:
 
 
 def send_message_to_discord_channel(
-    channel_id: int, message: dict, embed: bool = False
+    webhook, message: dict, embed: bool = False
 ) -> None:
 
-    # Check if the discordproxy module is active. We will use it as our priority app for notifications
+    logger.debug("Notifications: Starting to parse channel webhook notification")
+
+    url = webhook  # webhook url, from here: https://i.imgur.com/f9XnAew.png
+
+    # for all params, see https://discordapp.com/developers/docs/resources/webhook#execute-webhook
+    data = {}
+
+    # leave this out if you dont want an embed
+    # for all params, see https://discordapp.com/developers/docs/resources/channel#embed-object
+    data["embeds"] = [
+        {
+            "description": message["description"],
+            "title": message["title"],
+            "color": message["color"],
+            "footer": {"text": message["footer"]},
+            "fields": [
+                {"name": "Value", "value": message["value"], "inline": True},
+                {
+                    "name": "Assigned to",
+                    "value": message["assigned_to"],
+                    "inline": True,
+                },
+                {
+                    "name": "Assigned from",
+                    "value": message["assigned_from"],
+                    "inline": True,
+                },
+            ],
+        }
+    ]
+
+    result = requests.post(url, json=data)
+
     try:
-
-        from discordproxy.client import DiscordClient
-        from discordproxy.discord_api_pb2 import Embed
-        from discordproxy.exceptions import DiscordProxyException, DiscordProxyGrpcError
-
-        logger.debug("Attempting send message to discord channel %s" % channel_id)
-
-        client = DiscordClient()
-
-        fields = []
-
-        fields.append(Embed.Field(name="Value", value=message["value"], inline=True))
-        fields.append(
-            Embed.Field(name="Assigned to", value=message["assigned_to"], inline=True)
-        )
-        fields.append(
-            Embed.Field(
-                name="Assigned from", value=message["assigned_from"], inline=True
-            )
-        )
-
-        embed = Embed(
-            description=message["description"],
-            title=message["title"],
-            color=message["color"],
-            footer=Embed.Footer(text=message["footer"]),
-            fields=fields,
-            author=Embed.Author(name="AA Buyback Program"),
-        )
-
-        try:
-            logger.debug("Sending notification for discord channel %s" % channel_id)
-            client.create_channel_message(channel_id=channel_id, embed=embed)
-        except DiscordProxyGrpcError:
-            logger.debug(
-                "Discordprox is installed but not running, failed to send message. Attempting to send via aa-discordbot instead."
-            )
-            send_aa_discordbot_channel_notification(channel_id, message["description"])
-
-        except DiscordProxyException as ex:
-            logger.error("An error occured when trying to create a message: %s" % ex)
-
-    except ModuleNotFoundError:
-        logger.debug("Discordproxy is not installed, sending message via aa-discordbot")
-
-        send_aa_discordbot_channel_notification(channel_id, message["description"])
+        result.raise_for_status()
+    except requests.exceptions.HTTPError as err:
+        print(err)
+    else:
+        print("Payload delivered successfully, code {}.".format(result.status_code))